[ADVISOR-2338] Create systems table
This PR adds a modal when you click the Run Task button from the "Available tasks" tab on the tasks page. This modal should have details about the task and a system table.

This PR does not include the toolbar (other than the pagination) because tasks api does not support filtering and sorting yet. The code for filters is commented out for now and will be switched over when supported in the api.