[ADVISOR-2584] Execute an available task
This PR allows a user to execute an available task on selected systems.
This functionality will only work on a host that is turned on and actively connected. I'm not sure how to do this locally, but it can be done. Let me know if you test this and we can get a system actively connected and hard-code the id into the code to test it.

A successfully executed task should show a notification with a link to the task. An error should result in a notification noting the error.