[ADVISOR-2588] Populate run task modal with selected systems
This PR populates the run task modal with selected systems if they are passed in. This is for when you go to an executed task and click the "Run task again" button in the header. It will pass the systems that are in the jobs table through the component and pre-select them for running again. A user may still deselect and select systems other than those pre-populated.

At this point, the easiest way to test is to return an hard-coded id from the `getSelectedSystems` function in `CompletedTaskDetails` component; one that is in the list of systems and check to make sure it's selected when you load the modal.