[ADVISOR-2585] Delete executed task
This PR adds the ability to delete/cancel an executed (or in progress) task via Completed Tasks Table (https://marvelapp.com/prototype/6da50b0/screen/84773746) and via Complete Tasks Details (https://marvelapp.com/prototype/6da50b0/screen/84773747).

Deleting a task via the Completed Tasks Table should cause a loading state on the table and a re-render if the delete is successful. An unsuccessful delete results in toast alert.

Deleting a task via the Completed Tasks Details should cause the user to go back to the Completed Tasks Table if the delete is successful. An unsuccessful delete results in toast alert.

I have some commented out code for cancelling because it's not currently supported, but we plan to have support. I don't want to rewrite the code later, so I commented it out so I can turn it on later.

Make sure to test this on completed tasks and running tasks. In order to test a running task, make sure to run the "Run the insights-client" task and that should give you enough time before it completes. Let me know if you need to know which system to run it on.