[ADVISOR-2647] Add run task again to completd kebab
Add "Run task again" option to kebab on Completed Tasks table. This required some refactoring for the other instances of the RunTaskModal.

To test, you'll need to check the functionality of the RunTaskModal via
the following 3 locations:

Available Tasks
Completed Tasks table - kebab
Completed Task Details

Make sure to try to:
Run the task again
Add systems and cancel - reopening should still show original selections
Remove systems and cancel - reopening should still show original selections
Any other ways you can think to break it.

You may occassionally get an error resulting from trying to load task job details on a system that has since been deleted. This is a bug we are aware of. It is a good test of how the modal handles the errors though, so use it to your advantage.