[ADVISOR-2670] Add cancel task
This PR allows a task, that has not been completed, to be cancelled. If a task status is "Running", the kebab dropdown on the completed tasks table should include "Cancel" instead of delete. The same is true if you view the details page of a completed task. The kebab in the right of the top header should include "Cancel" instead of delete.

After clicking cancel, there are a number of things you will see in the modal. You will have the option to: Cancel a task - this will simply cancel but leave all details of the completed task in the table Cancel and delete a task - this will remove the task from your table and the backend will handle cancelling and deleting.

Additionally, a task can only cancel jobs on systems that are connected via Satellite. A system with any other means of connection cannot be cancelled. Therefore, when entering the cancel modal, if any system is NOT connected via Satellite, there will be a sentence in the modal body stating, "Tasks running on systems connected with rhc cannot be cancelled." Likewise, if a task is running on systems ONLY connected via Satellite, that sentence will not be visible.

Now, at this point, cancelling isn't working properly. The backend is not able to properly cancel a task. So, when you cancel a task, you'll notice the task returns a status of "Cancelled", but the jobs will still say they are "Running". This much is normal. What we would expect is to eventually see those jobs return as "Cancelled" if they are connected via Satellite, and some other status if they are connected via rhc. But what currently happens is all systems finish their runs and the status of the task completes and the "Run end" value includes a successful run completion time.

The frontend code should work fine when the backend code is implemented.