[ADVISOR-2705] Disable delete when task is running
Before, if a task was running, you could delete it during the run. Now, we don't want to allow a deletion until the run is complete.

To test, run a task (preferrably the "Run the insights-client", and then try to delete that task via the Executed tasks table kebab and on the task details page in the top right kebab.

With this fix, it should be disabled.