[ADVISOR-2710] Fix empty state bugs
In NoResultsTable.js in emptyRows function you're hardcoding colspan: 3 which does not cause any immediate problems, but imagine you added new columns to table that's using it or used it for a larger table, then the empty state would be off center, you can either pass exact amount of columns, or simply set the colspan to 100, so you virtually never run out, we use second option in both Vulnerabilities and it works wonders.

If a table items are selectable with checkboxes, then empty state has checkbox as well. To test, go to systems table and enter a filter that causes no results. You'll see a checkbox on the empty state.

These two issues are fixed in this PR.