Fix run task button issue
Before, if you clicked on available task tab and click the "Run this task" button on a task, it will error out. Now, this button should open the modal and display the task information and system table.

This action should be the same for the "Run this task again" button via the completed task details page.