[ADVISOR-2824] Add expandable rows on jobs table
When a task has completed its run if a job has completed with a status of "Success", but that job has issues, you will get an option for an expandable row for that job only. If there are no jobs with a status of "Success" that have issues, then you will see no expandable row carats.

When you expand the row, you should get a variable number of rows that highlight the issues for those jobs. They should be sorted with highest risk issue first. Upon expanding these rows you should find additional details that include a summary of the issue, any remediations (not for info items), and a key.

Let me know if you need to find a system to run a task on that will give you a success with issues.