[ADVISOR-2864] Fix deleted system in Run task again
To repro:
Find a completed task with a deleted system
Click the Run task again button in the top right
Click the Execute task button

Notice when you load the systems table, the count in the bulk select includes the deleted system, but the actual number of selected items is lower than that number.

When you click to execute the task, you will get an error because you submitted an id that no longer exists.

This PR fixes both of these issues.