Fix expandable bug
Before, if a task completed successfully and had a report, but alert: false, the details wouldn't render properly and the expandable row wasn't present. This PR fixes that.

This PR also fixes a sort issue on the Activity table "Systems" row.

This PR also fixes a bug to render the icon in the labels for task details.