[ADVISOR-3090] Display text when no report_json
Currently, we have only one task that displays job results in the expanded row. That content is unique to the specific task and we need to make a failsafe so that if a job is run that doesn't have a `report_json` that it will fallback to the `report` property and display the text.

To test, go to this line:
https://github.com/RedHatInsights/tasks-frontend/blob/master/src/SmartComponents/CompletedTaskDetails/CompletedTaskDetails.js#L76
and replace it with
https://github.com/RedHatInsights/tasks-frontend/blob/master/src/SmartComponents/CompletedTaskDetails/__tests__/__fixtures__/completedTasksDetails.fixtures.js#L62.

This will use the fixture to display the scenario we're testing. In the instance where `report_json` is not provided, we'll rely on the backend to provide text in the `report` property of the object.

Now, go to Automation Toolkit -> Tasks and click the Activity tab. Select any task, as the mocked data will replace any actual data.

Expand the bottom job and compare it to the other expandable jobs to see the difference.