feat(HMS-1925):Add Centos preconversion support
This PR mainly makes some components more generic so different types of tasks can display on the same component. We've done this by creating constants that define how each unique task defines its statuses and other data, and then we map the executed task to the constant via the task_slug and pull in the constant.