fix(HMS-2644): catch new remediation field structure
When a pre-upgrade analysis task has a job that doesn't return a remediation, then it leaves the field off completely. We currently handle that correctly. But the new pre-conversion analysis tasks have jobs that, when a remediation isn't provided, still provides the remediation key value pair, but it is empty.

This PR handles both scenarios. Let me know if you need ids to test both types of tasks.