fix: properly filter by os version
Currently in stage-beta, if you add os filters, the system fetch will break because it's not properly adding the versions to the request string. This PR properly creates the fetch request.

It will only work in stage-beta at this point. It is currently not fetching correctly because the backend needs a change. if you filter by centos version 7.9 and rhel version 8.8, it will try to fetch centos versions 7.9 and 8.8 and rhel versions 7.9 and 8.8.