fix(RHINENG-2529): show centos versions in os filter
With this PR, CentOS Linux versions should show up in the os filter dropdown on the inventory table. Go to Available tasks tab and click Run task on one of the available tasks. The systems table will populate there. In stage-preview, you shouldn't see the CentOS Linux versions, but this PR exposes them.