fix(RHELC-1186): diagnosis doesn't show for pre-conversion
We were incorrectly checking the `detail.diagnosis` value as an object when it was, in fact, an array. This PR combines some code for simplicity and properly checks the value and displays it.

To test, in insights account, check `executed_task/5330`. You should be able to expand some of the job details and see that the `diagnosis` field doesn't populate in the details. When running this PR, you should see it.

Please also check that `executed_task/5331` also works correctly by continuing to show `remediations` field.