fix(RHINENG-3105): loading rows not showing on activity delete
Description of Problem
There's a small bug in Tasks on the Activity table. When you're on any page other than 1 and you delete an activity, the table shows a No results empty state instead of loading rows.

How reproducible
Always

Steps to Reproduce
Go to Tasks
Click Activity Tab
Go to page 2+
Delete an activity via the kebab
Actual Behavior
Table shows a "No results" empty state.

Expected Behavior
Should show loading rows

Additional info
This happens because the table component receives an array of 5 loading row components. Then it tries to paginate that data, and the `page` is currently set to a page over page 1, so if you're looking at 10 per page and you're on page 2, the 5 items sent in get paginated out.

This fix will skip pagination if the table is in a loading state.