feat(RHIN-1101): Add Input Parameters to Tasks (#126)
* feat(RHIN-1101): Add Input Parameters to Tasks

Some tasks will request certain parameters in order to complete the task. Those parameters may or may not be required, but they will be a part of the response when you fetch task details.

When a task doesn't have parameters, you can execute the task after selecting systems. If a task does have parameters, you can click the "Next" button after selecting a system.

Clicking next button will bring up the input page for parameters. As long as all required parameters are filled in, then you can execute the task.

Required parameters are validated by the form input. On first load of the page, the validation error text will not show on an empty input. But if you type something and delete it, then you'll see the validation error appear. Either way, the Execute button should not be enable unless all required params are filled in.

To test, there is a "Test Task with parameters" task on the Available tab of the Tasks app. You can run that task in full to test. But, to test additional variations of possible parameters, replace the parameters in RunTaskModal with those from the fixtures.

You'll also need to test the functionality of "Run this task again". You can do that by running the "Test task with parameters" task and from both the kebab in the Activity Table and from the button in the top right of the completed task details, you should be able to generate a new task with the same parameter settings of the original task.

* Make review changes and add tests

---------

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>