fix(HMS-3139) Temporarily display overridable as inhibitor
This PR attempts to display overridable results in convert2rhel tasks as inhibitors. To test, login to insights account and go to `/tasks/executed/6354`. The first entry should be displayed as an inhibitor if you expand the row. If you check the `jobs` network request, you'll see that the entry is actually set to 'OVERRIDABLE' severity.

The script will be updated later for overriable tasks and we'll simply convert the `TaskEntries` back and update tests.