fix(RHINENG-7695): Change download playbook to script
"Looking at the pre-conversion for convert2RHEL, the link should say I can download the script.. not download the playbook."

This PR compares the task slug and when it's a conversion task, it will show the "Download preview..." link on the Available tasks page to "script" instead of "playbook". All other tasks will remain as "Download preview of playbook".