feat(RHINENG-2263): Add log viewer to task details (#155)
* feat(RHINENG-2263): Add log viewer to task details

When a task has completed, that task may have a job or series of jobs that have logs. This PR adds the ability for a user to view these logs in the UI.

A "View logs" link should appear when a job (system) is expanded and that job has logs. When the link is clicked, a drawer with a log viewer component is displayed. That log viewer has the ability to search, export as a file, and open the log in a new tab.

For now, we are adding the "View log" link in the expanded details of the job. In the next iteration, we will be adding the "View logs" action to a kebab in the table next to the job. For now, let's test the link.

Finding a task with logs may be difficult, but if you run this on prod.foo and sign into the insights-qa account, you can go to `/insights/tasks/executed/10753` and view a task with logs.

* Add tests

* Move drawer into Page component; fix tests

---------

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>