fix(RHINENG-10944): Move Drawer out of Page component
The JobLogDrawer has been moved out of the Page component and wrapped around the contents of the CompletedTaskDetails component.

Previously, we had to put the Drawer inside of the Page component in order to display the drawer correct when expanded. But this caused there to be an additional scrollbar on the page.

Steps to Reproduce
- Find a tasks execution with more than 20 hosts or execute one
- Click on the task details to see individual host results
- Select the paginator to show more than 10 (or simply shrink the height of the page)
- Try to scroll the list

Also, test the JobLogDrawer by clicking the kebab on a system and click "View system logs" and make sure the functionality has remained.